+++ tests/testthat/test-kaplan_meier_plot.R
@@ -9,7 +9,7 @@ get_test_data <- function() {
 }
 
 # h_xticks ----
-test_that("h_xticks works with xticks NULL", {
+test_that("h_xticks works with default settings", {
   result <- h_data_plot(get_test_data()) %>%
     h_xticks()
   expected <- seq(0, 5000, 1000)
@@ -18,7 +18,7 @@ test_that("h_xticks works with xticks NULL", {
 
 test_that("h_xticks works with xticks number", {
   result <- h_data_plot(get_test_data()) %>%
-    h_xticks(100)
+    h_xticks(xticks = 100)
   expected <- seq(0, 4700, 100)
   expect_identical(result, expected)
 })
@@ -26,9 +26,31 @@ test_that("h_xticks works with xticks number", {
 test_that("h_xticks works with xticks numeric", {
   expected <- c(0, 365, 1000)
   result <- h_data_plot(get_test_data()) %>%
-    h_xticks(expected)
+    h_xticks(xticks = expected)
   expect_identical(result, expected)
 })
+
+test_that("h_xticks works with max_time only", {
+  expected <- c(0, 1000, 2000, 3000)
+  result <- h_data_plot(get_test_data(), max_time = 3000) %>%
+    h_xticks(max_time = 3000)
+  expect_identical(result, expected)
+})
+
+test_that("h_xticks works with xticks numeric when max_time is not NULL", {
+  expected <- c(0, 365, 1000)
+  result <- h_data_plot(get_test_data(), max_time = 1500) %>%
+    h_xticks(xticks = expected, max_time = 1500)
+  expect_identical(result, expected)
+})
+
+test_that("h_xticks works with xticks number when max_time is not NULL", {
+  expected <- c(0, 500, 1000, 1500)
+  result <- h_data_plot(get_test_data(), max_time = 1500) %>%
+    h_xticks(xticks = 500, max_time = 1500)
+  expect_identical(result, expected)
+})
+
 
 # h_tbl_median_surv ----
 test_that("h_tbl_median_surv estimates median survival time with CI", {